Moved active filters above points and filter columns.
