+++ styles/FilterList.module.css
@@ -16,6 +16,8 @@
     padding: 0;
     list-style-type: none;
     margin: 0;
+    overflow-y: scroll;
+    max-height: 110px; 
 }
 
 .activeFilterItem {
